refactor(react-architecture): rename asyncComponent to async
