+++ src/async/asyncAdminDashboard.js
@@ -1,7 +1,7 @@
 import withAsyncComponent from '../hoc/withAsyncComponent';
 
 const asyncAdminDashboard = withAsyncComponent(() => {
-  return import('../applications/AdminDashboard')
+  return import('../containers/AdminDashboard')
 })
 
 export default asyncAdminDashboard
