+++ src/async/asyncLogout.js
