--- src/asyncComponents/asyncAdminDashboard.js
@@ -1,7 +1,0 @@
-import withAsyncComponent from '../hoc/withAsyncComponent';
-
-const asyncAdminDashboard = withAsyncComponent(() => {
-  return import('../containers/AdminDashboard')
-})
-
-export default asyncAdminDashboard
