+++ src/containers/ForgotPassword/actions.js
@@ -1,0 +1,7 @@
+export const handleRecover = (ctx, event) => {
+  event.preventDefault()
+  ctx.setState({
+      isRecoverSent: true
+  })
+  ctx.props.actions.fetchRecoverPassword()
+}
